Updated dependencies
And ported over all deprecated passes and test fixtures.
