+++ src/main/scala/io/shiftleft/js2cpg/cpg/passes/BuiltinTypesPass.scala
@@ -4,10 +4,10 @@ import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.codepropertygraph.generated.nodes.{NewNamespaceBlock, NewType, NewTypeDecl}
 import io.shiftleft.codepropertygraph.generated.{EdgeTypes, NodeTypes}
 import io.shiftleft.js2cpg.cpg.datastructures.OrderTracker
-import io.shiftleft.passes.SimpleCpgPass
+import io.shiftleft.passes.CpgPass
 import org.slf4j.LoggerFactory
 
-class BuiltinTypesPass(cpg: Cpg) extends SimpleCpgPass(cpg) {
+class BuiltinTypesPass(cpg: Cpg) extends CpgPass(cpg) {
 
   private val logger = LoggerFactory.getLogger(getClass)
